requirements.txt
pyspark==3.3.1
faker
delta-spark
pylint
black

pytest